Minor correction to test assertion for Attack()
plasma chaingun damage increases as target gets farther away so it rounds to 0 when distance is only 1.  Moved target to get a bigger value

+++ src/weapon/ChainGun.java
@@ -42,7 +42,7 @@ public class ChainGun extends GenericWeapon
 	public int damageCalculation(int distance)
 	{
 
-		return (int) ((int) (float) baseDamage * (((float) (distance) / maxRange)));
+		return (int) (baseDamage * (((float) (distance) / maxRange)));
 
 	}
